Caching removed

+++ backend/src/main/java/hu/exyxwd/tisztatisza/controller/UpdateLogController.java
@@ -7,7 +7,9 @@ import lombok.AllArgsConstructor;
 
 import hu.exyxwd.tisztatisza.model.UpdateLog;
 import hu.exyxwd.tisztatisza.repository.UpdateLogRepository;
+import hu.exyxwd.tisztatisza.exception.ResourceNotFoundException;
 
+// TODO: response entities
 @RestController
 @AllArgsConstructor
 @RequestMapping("/api/logs")
@@ -21,6 +23,11 @@ public class UpdateLogController {
 
     @DeleteMapping
     public void deleteLogs(@RequestBody List<Long> ids) {
+        List<UpdateLog> logs = updateLogRepository.findAllById(ids);
+        if (logs.size() != ids.size()) {
+            throw new ResourceNotFoundException("Some logs do not exist with the given ids");
+        }
+
         updateLogRepository.deleteAllById(ids);
     }
 }
